ClientDB. Комплектация ИБ по системам. Очередные исправления

+++ ClientDB/StoredProcedures/USR.COMPLECT_INFO_BANK_CACHE_RESET.sql
@@ -6,22 +6,35 @@ SET QUOTED_IDENTIFIER ON
 GO
 
 CREATE PROCEDURE [USR].[COMPLECT_INFO_BANK_CACHE_RESET]
+	@Complect	VarChar(100) = NULL
 AS
 BEGIN
 	SET NOCOUNT ON;
 	
-	TRUNCATE TABLE dbo.ComplectInfoBankCache
+	IF @Complect IS NULL
+		TRUNCATE TABLE dbo.ComplectInfoBankCache
+	ELSE
+		DELETE FROM dbo.ComplectInfoBankCache WHERE Complect = @Complect;
 	
 	INSERT INTO dbo.ComplectInfoBankCache(Complect, InfoBankID, InfoBankName)
-	SELECT DISTINCT rns.Complect, cgl.InfoBankID, cgl.InfoBankName
-	FROM Reg.RegNodeSearchView rns WITH(NOEXPAND)
+	SELECT rns.Complect, cgl.InfoBankID, cgl.InfoBankName
+	FROM
+	(
+		SELECT DISTINCT rns.Complect
+		FROM Reg.RegNodeSearchView rns WITH(NOEXPAND)
+		WHERE	DS_REG = 0
+			AND SubhostName NOT IN ('Ó1', 'Í1', 'Ì', 'Ë1')
+			AND (rns.Complect = @Complect OR @Complect IS NULL)
+	) rns
 	CROSS APPLY dbo.ComplectGetBanks(rns.Complect, NULL) cgl
-	WHERE DS_REG = 0 AND SubhostName NOT IN ('Ó1', 'Í1', 'Ì', 'Ë1') AND 
-			NOT EXISTS
-				(
-					SELECT *
-					FROM Reg.RegNodeSearchView rns2
-					WHERE	rns2.Complect = rns.Complect AND
-							SubhostName IN ('Ó1', 'Í1', 'Ì', 'Ë1')
-				)
+	--/*
+	WHERE NOT EXISTS
+		(
+			SELECT *
+			FROM Reg.RegNodeSearchView rns2 WITH(NOEXPAND)
+			WHERE	rns2.Complect = rns.Complect
+				AND SubhostName IN ('Ó1', 'Í1', 'Ì', 'Ë1')
+				AND DS_REG = 0
+		)
+	--*/
 END
